pc - update for storybook 8
