--- SAVE-STORIES/pages/HomePage/SectionSearches/SectionSearchesIndexPage.stories.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-import SectionSearchesIndexpage from "main/pages/SectionSearches/SectionSearchesIndexPage";
-
-export default {
-  title: "pages/SectionSearches/SectionSearchesIndexPage",
-  component: SectionSearchesIndexpage,
-};
-
-const Template = () => <SectionSearchesIndexpage />;
-
-export const Default = Template.bind({});
